+++ pytsite/admin/_navbar.py
@@ -1,4 +1,3 @@
 __author__ = 'Alexander Shepetko'
 __email__ = 'a@shepetko.com'
 __license__ = 'MIT'
-