+++ pytsite/content/_widget.py
@@ -5,7 +5,7 @@ __email__ = 'a@shepetko.com'
 __license__ = 'MIT'
 
 from pytsite import taxonomy as _taxonomy, auth as _auth
-from pytsite.core import widget as _widget, html as _html, lang as _lang
+from pytsite.core import widget as _widget, html as _html, lang as _lang, router as _router
 from . import _model, _functions
 
 
@@ -22,7 +22,7 @@ class ContentModelSelect(_widget.select.Select):
 
 
 class TagCloud(_taxonomy.widget.Cloud):
-    """Tags Clod Widget.
+    """Tags Cloud Widget.
     """
     def __init__(self, limit=10, **kwargs):
         """Init.
@@ -31,7 +31,7 @@ class TagCloud(_taxonomy.widget.Cloud):
 
 
 class EntityTagCloud(_widget.Base):
-    """Tag of the Entity Widget.
+    """Tags Cloud of the Entity Widget.
     """
     def __init__(self, entity: _model.Content, **kwargs):
         """Init.
@@ -55,3 +55,22 @@ class EntityTagCloud(_widget.Base):
             root.append(a)
 
         return self._group_wrap(root)
+
+
+class SearchInput(_widget.Base):
+    def __init__(self, model: str, **kwargs):
+        super().__init__(**kwargs)
+        self._model = model
+
+    def render(self) -> _html.Element:
+        form = _html.Form(cls='wrapper form-inline', method='GET')
+        form.append(_html.Input(type='text', cls='form-control', name='search',  required=True,
+                                placeholder=_lang.t('pytsite.content@search_input_placeholder'),
+                                value=_router.request.values_dict.get('search', '')))
+        form.set_attr('action', _router.endpoint_url('pytsite.content.eps.index', {
+            'model': self._model,
+        }))
+
+        btn = _html.Button(type='submit', cls='btn btn-default')
+        form.append(btn.append(_html.I(cls='fa fa-search')))
+        return self._group_wrap(form, False)